+++ CHANGELOG.md
@@ -9,6 +9,10 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Added
+
+- Add `path()` method to `ASTCache` class https://github.com/nuwave/lighthouse/pull/2694
+
 ## v6.58.0
 
 ### Changed